QCamera2: stack: Increase camera open retry count
If camera daemon crashes, it could take up to 10 seconds for it
to restarts. This time seems to increase in latest OS upgrade.

Increasing the retry count makes it less likely to fail get_camera_info.

Test: Hardcode hasPendingBuffers in HAL destructor to force restart
Bug: 67576800

+++ camera/QCamera2/stack/mm-camera-interface/src/mm_camera.c
@@ -238,7 +238,7 @@ int32_t mm_camera_open(mm_camera_obj_t *my_obj)
 {
     char dev_name[MM_CAMERA_DEV_NAME_LEN];
     int32_t rc = 0;
-    int8_t n_try=MM_CAMERA_DEV_OPEN_TRIES;
+    int32_t n_try=MM_CAMERA_DEV_OPEN_TRIES;
     uint8_t sleep_msec=MM_CAMERA_DEV_OPEN_RETRY_SLEEP;
     unsigned int cam_idx = 0;
